reorganisation du main + ajout fichier UI
